// src/shared/enums/product.ts
export enum EProductAssignment {
  FOR_AIRPODS = 'for_airpods',
  FOR_CARDS = 'for_cards',
  FOR_CASH = 'for_cash',
  FOR_CLOTH = 'for_cloth',
  FOR_COINS = 'for_coins',
  FOR_DOCUMENTS = 'for_documents',
  FOR_WATCH = 'for_watch',
}
export enum EProductCategory {
  BELT = 'belt',
  BIFOLD_WALLET = 'bifold_wallet',
  CARD_HOLDER = 'card_holder',
  DOC_HOLDER = 'doc_holder',
  HOLDER_FOR_CAR_DOCS = 'holder_for_car_docs',
  PASSPORT_COVER = 'passport_cover',
  PURSE = 'purse',
  WATCH_STRAP = 'watch_strap',
}